requirements.txt
cairosvg
Pillow
numpy
pytoshop
